refactor db files, and add error handling to db connection
